use shared implementation of getting github auth url
