+++ frontend/src/user-info-bar/enable-private-repos-button.tsx
@@ -1,27 +1,19 @@
 import React, { FunctionComponent } from 'react';
 import { Button } from '@material-ui/core';
 
-import { config } from '../config';
+import { getGithubAuthenticationUrl } from '../authentication/get-github-authentication-url';
+import { RepositoriesPermissions } from '../authentication/repositories-permissions';
 
 export const EnablePrivateReposButton: FunctionComponent = () => {
   return (
     <Button
       variant="contained"
       color="primary"
-      href={getGithubAuthenticationUrl()}
+      href={getGithubAuthenticationUrl(
+        RepositoriesPermissions.PrivateAndPublic,
+      )}
     >
       Enable private repos
     </Button>
   );
-};
-
-const getGithubAuthenticationUrl = () => {
-  const redirectUrl = window.location.href;
-  const queryParams = [
-    `client_id=${config.githubClientID}`,
-    `redirect_uri=${redirectUrl}`,
-    'scope=repo',
-  ].join('&');
-
-  return `https://github.com/login/oauth/authorize?${queryParams}`;
 };
